fix: change calculation of overall score

+++ app/src/main/java/com/rollcloud/doon/utils.kt
@@ -2,7 +2,7 @@ package com.rollcloud.doon
 
 import kotlin.math.*
 
-fun Double.toHoursAndDays(signed: Boolean = false): String {
+fun Float.toHoursAndDays(signed: Boolean = false): String {
   val sign = this.sign
   val whole = floor(this.absoluteValue).toInt()
   val decimal = this.absoluteValue - whole
